+++ .idea/misc.xml
@@ -43,6 +43,11 @@
         <entry key="../../../../../layout/compose-model-1655256856215.xml" value="0.1" />
         <entry key="../../../../../layout/compose-model-1655256872738.xml" value="0.38333333333333336" />
         <entry key="../../../../../layout/compose-model-1655259203218.xml" value="0.5017482517482518" />
+        <entry key="../../../../../layout/compose-model-1659761295695.xml" value="0.19552364864864866" />
+        <entry key="../../../../../layout/compose-model-1659761304495.xml" value="0.1" />
+        <entry key="../../../../../layout/compose-model-1659761312617.xml" value="0.1" />
+        <entry key="../../../../../layout/compose-model-1659761319450.xml" value="0.1" />
+        <entry key="../../../../../layout/compose-model-1659761327981.xml" value="1.2074468085106382" />
         <entry key="..\:/Users/CarlosMenjivar/Documents/repositories/pomodoro-scheduler/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.17" />
         <entry key="..\:/Users/CarlosMenjivar/Documents/repositories/pomodoro-scheduler/app/src/main/res/drawable/ic_launcher.xml" value="0.17" />
         <entry key="..\:/Users/CarlosMenjivar/Documents/repositories/pomodoro-scheduler/app/src/main/res/drawable/ic_launcher_background.xml" value="0.127" />
