--- app/src/main/java/com/emenjivar/pomodoro/components/Item.kt
@@ -1,45 +1,0 @@
-package com.emenjivar.pomodoro.components
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun Item(
-    modifier: Modifier = Modifier,
-    title: String,
-    subtitle: String? = null,
-    onClick: () -> Unit,
-    rightContent: @Composable () -> Unit
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .clickable { onClick() }
-            .padding(16.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        Column(modifier = Modifier.weight(2f)) {
-            Text(text = title)
-            if (subtitle != null) {
-                Text(text = subtitle, style = MaterialTheme.typography.caption)
-            }
-        }
-
-        Column(
-            modifier = Modifier.weight(0.8f),
-            horizontalAlignment = Alignment.End,
-            content = { rightContent() }
-        )
-    }
-}
