+++ app/src/main/java/com/emenjivar/pomodoro/system/CustomNotificationManagerImp.kt
@@ -9,8 +9,8 @@ import android.os.Build
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 import com.emenjivar.pomodoro.R
-import com.emenjivar.pomodoro.model.Counter
-import com.emenjivar.pomodoro.model.Phase
+import com.emenjivar.pomodoro.utils.model.Counter
+import com.emenjivar.pomodoro.utils.model.Phase
 import com.emenjivar.pomodoro.utils.Action
 import com.emenjivar.pomodoro.utils.formatTime
